Ported more tests, fixed an exception safety bug.
Summary:
Ported the test_cuckoo.cc test and the test_iterator.cc test. The
exception safety bug was that we were doing multiple allocations in
the constructor of the cuckoohash_map. If one allocation succeeded and
the next failed, there would be a memory leak. I wrapped the pointers
in std::unique_ptr's, to enforce singular ownership of the pointer and
so that the pointer would be freed upon calling it's destructor.

Test Plan: testing/

Reviewers: binfan

CC: manugoyal, binfan, dga, kaminsky

Differential Revision: https://sourcery.cmcl.cs.cmu.edu:8080/D1