Fixed (mostly) the issue with freeing memory after expansion
Summary:
Whereas before expansion would wait a certain amount of time before
freeing the old memory for the table, it now stores that pointer and
leaves it valid until the next expansion, during which it frees the
stored pointer and stores the next old pointer. This isn't 100% safe,
but the only case in which it would cause a problem is if two expansions
run in the middle of a find, which is unlikely to happen.

Test Plan: testing/

Reviewers: binfan, dga, kaminsky

Differential Revision: https://sourcery.cmcl.cs.cmu.edu:8080/D4