Integrated fine-grained locking
For now, locks replace the versions array, so finds contend with the
other operations. This greatly improves insert and delete performance,
but slightly impacts finds. Also instead of deleting old memory when
expanding, we store it in a list and delete all the memory when the hash
table destructor is called.