Made libcuckoo more platform-independent
If its a linux machine, we use sched_getcpu to pick the counter to
increment, and we return 0 otherwise. Also gtest was being annoying to
compile on mac, so I removed it entirely and rewrote the macros I
needed, since we weren't using it much anyways. Also other minor
refactors.