Miscellaneous refactors and bug fixes
Added a template specialization for CityHasher for strings, introduced
an expansion_lock to serialize expansions and iterators, and fixed a bug
with update. Also added a template parameter for the equality function,
since the STL does so. Also moved the spinlock and iterator classes
inside the cuckoohash_map definition to enforce proper encapsulation.