Added another stress test plus refactors
The original stress test only looked at inserts, deletes, updates, and
finds, and checked everything for correctness. This one tests every
public table function, including resizing and iterators along with the
others, but doesn't check for correctness. Also fixed small things and
made the table more STL-like. There was a weird bug with the new clang
that came with Mavericks that created errors for file-scoped atomic
variables in the tests, so I moved some of them inside other stuff.